Limit the operation type for scheduledoperation
Now karbor only support two types of scheduledoperation,
but the client do not show and limit the values. So the
end users can type any string to execute the command of
'scheduledoperation create', but the server returns error
, this will make users confused, and they still not know
the right value. This patch will limit the operation type
in 'protect' and 'retention_protect'.

Change-Id: Ic1110124472ac455f988bb25254feeb4417caf1a