Fix listing with --all error
When executing command "karbor xxx-list --all", it will raise
error as 'error: ambiguous option: --all could match --all-tenants,
--all_tenants'. The reason is we have both '--all-tenants' and
'--all_tenants' in the specify operations, '--all' matches two
args, so it can not work, but when using '--all-' or '--all_',
it return the correct result. We should fix it, so we remove the
arg '--all_tenants' which is not in the help info.
Story: 2005874
Task: 33686

Change-Id: Iafa70c35594af732435122ebd50c114fd7f0b9df